# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Quiz App",
        "description": "Documentación de mi API REST",
        "contact": {
            "email": "kvfa13@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/quiz/{quizId}/questions": {
            "get": {
                "tags": [
                    "Questions"
                ],
                "summary": "Listar perguntas por quiz",
                "description": "Retorna uma lista de perguntas associadas a um quiz específico.",
                "operationId": "e028629a9a0f98cc7b6360e17dce1aa7",
                "parameters": [
                    {
                        "name": "quizId",
                        "in": "path",
                        "description": "ID do quiz para o qual as perguntas devem ser listadas",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de perguntas do quiz",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 19
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Quiz 2"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Queremos eliminar um quiz de forma correta"
                                                },
                                                "questions": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 25
                                                            },
                                                            "question": {
                                                                "type": "string",
                                                                "example": "hola mundo"
                                                            },
                                                            "image": {
                                                                "type": "string",
                                                                "example": null,
                                                                "nullable": true
                                                            },
                                                            "correct_answer": {
                                                                "type": "integer",
                                                                "example": 0
                                                            },
                                                            "answer_count": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "quiz_name": {
                                                                "type": "string",
                                                                "example": "Quiz 2"
                                                            },
                                                            "quiz_id": {
                                                                "type": "integer",
                                                                "example": 19
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Quiz não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No results found for Quiz with ID 18"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/questions/{id}": {
            "delete": {
                "tags": [
                    "Questions"
                ],
                "summary": "Deletar uma pergunta",
                "description": "Permite que apenas administradores deletam uma pergunta específica. Esta ação é restrita a usuários com o papel de administrador.",
                "operationId": "2cde4cbb83a9fc1d3660444e77750195",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da pergunta que será deletada.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pergunta deletada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 19
                                        },
                                        "quiz_id": {
                                            "type": "integer",
                                            "example": 18
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "question": {
                                            "type": "string",
                                            "example": "amas a Dios ?"
                                        },
                                        "image": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "correct_answer": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-10-29T03:42:21.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-10-29T03:42:21.000000Z"
                                        },
                                        "quiz": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 18
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Quiz 1"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Esto es un ejemplo de quiz"
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-10-29T01:58:21.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-10-29T01:58:21.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pergunta não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No results found for Question with ID 19"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}